Creado y movido el detalle oferta a su controlador (#173)
